# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeriPulse Trust Intelligence API",
    "description": "Enterprise-grade Trust Intelligence combining email deliverability, disposable domain detection, phone carrier intelligence, and VoIP fraud scoring into a unified 0-100 Trust Score.",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://raw.githubusercontent.com/jasdebarreau3-cell/veripulse/main/assets/veripulse_logo.png",
      "altText": "VeriPulse Labs Logo"
    }
  },
  "paths": {
...
    "/v1/trust-score": {
      "get": {
        "tags": [
          "Trust Intelligence"
        ],
        "summary": "Evaluate Trust Score",
        "description": "Unified Signup Protection & Fraud Scoring Engine.\nCombines email deliverability, disposable domain checks, phone line intelligence,\nand VoIP carrier detection into a single, definitive 0-100 Trust Score.",
        "operationId": "evaluate_trust_score_v1_trust_score_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Email address to evaluate",
              "title": "Email"
            },
            "description": "Email address to evaluate"
          },
          {
            "name": "phone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Phone number to evaluate",
              "title": "Phone"
            },
            "description": "Phone number to evaluate"
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default ISO country code for phone number",
              "default": "US",
              "title": "Country Code"
            },
            "description": "Default ISO country code for phone number"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TrustScoreRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trust Intelligence"
        ],
        "summary": "Evaluate Trust Score",
        "description": "Unified Signup Protection & Fraud Scoring Engine.\nCombines email deliverability, disposable domain checks, phone line intelligence,\nand VoIP carrier detection into a single, definitive 0-100 Trust Score.",
        "operationId": "evaluate_trust_score_v1_trust_score_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Email address to evaluate",
              "title": "Email"
            },
            "description": "Email address to evaluate"
          },
          {
            "name": "phone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Phone number to evaluate",
              "title": "Phone"
            },
            "description": "Phone number to evaluate"
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default ISO country code for phone number",
              "default": "US",
              "title": "Country Code"
            },
            "description": "Default ISO country code for phone number"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TrustScoreRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verify": {
      "get": {
        "tags": [
          "Email Intelligence"
        ],
        "summary": "Verify Email",
        "operationId": "verify_email_v1_verify_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Email address to evaluate",
              "title": "Email"
            },
            "description": "Email address to evaluate"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verify-batch": {
      "post": {
        "tags": [
          "Email Intelligence"
        ],
        "summary": "Verify Batch",
        "operationId": "verify_batch_v1_verify_batch_post",
        "parameters": [
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/phone/validate": {
      "get": {
        "tags": [
          "Phone Intelligence"
        ],
        "summary": "Validate Phone",
        "operationId": "validate_phone_v1_phone_validate_get",
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Phone number to evaluate (e.g. +14155552671)",
              "title": "Phone"
            },
            "description": "Phone number to evaluate (e.g. +14155552671)"
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default ISO country code",
              "default": "US",
              "title": "Country Code"
            },
            "description": "Default ISO country code"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchVerificationRequest": {
        "properties": {
          "emails": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Emails",
            "description": "List of emails to verify (max 50 per batch)"
          }
        },
        "type": "object",
        "required": [
          "emails"
        ],
        "title": "BatchVerificationRequest"
      },
      "BatchVerificationResponse": {
        "properties": {
          "total_processed": {
            "type": "integer",
            "title": "Total Processed",
            "description": "Number of emails evaluated in the batch"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/EmailVerificationResponse"
            },
            "type": "array",
            "title": "Results",
            "description": "Individual verification results"
          }
        },
        "type": "object",
        "required": [
          "total_processed",
          "results"
        ],
        "title": "BatchVerificationResponse"
      },
...
      "EmailVerificationResponse": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "The queried email address"
          },
          "normalized_email": {
            "type": "string",
            "title": "Normalized Email",
            "description": "Standardized lowercase email address"
          },
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid",
            "description": "Overall deliverability status (syntax valid & active mail server)"
          },
          "is_syntax_valid": {
            "type": "boolean",
            "title": "Is Syntax Valid",
            "description": "Conforms to standard RFC email formatting"
          },
          "is_disposable": {
            "type": "boolean",
            "title": "Is Disposable",
            "description": "True if domain is a known burner/disposable email provider"
          },
          "has_mx_records": {
            "type": "boolean",
            "title": "Has Mx Records",
            "description": "True if domain has active mail routing (MX or A fallback)"
          },
          "is_free_provider": {
            "type": "boolean",
            "title": "Is Free Provider",
            "description": "True if provided by consumer webmail (Gmail, Yahoo, Outlook, etc.)"
          },
          "is_role_account": {
            "type": "boolean",
            "title": "Is Role Account",
            "description": "True if account is generic alias (e.g., admin, support, sales)"
          },
          "did_you_mean": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Did You Mean",
            "description": "Suggested domain correction if a common typo was detected"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "description": "Risk rating from 0 (safe) to 100 (high risk/fraudulent)"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "Risk tier: LOW, MEDIUM, HIGH, or CRITICAL"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action",
            "description": "Recommended signup action: ALLOW, FLAG_FOR_REVIEW, or BLOCK"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons",
            "description": "Explanatory signals contributing to the score"
          },
          "mx_records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Mx Records",
            "description": "Discovered mail exchange records"
          }
        },
        "type": "object",
        "required": [
          "email",
          "normalized_email",
          "is_valid",
          "is_syntax_valid",
          "is_disposable",
          "has_mx_records",
          "is_free_provider",
          "is_role_account",
          "risk_score",
          "risk_level",
          "recommended_action"
        ],
        "title": "EmailVerificationResponse"
      },
      "FormatsInfo": {
        "properties": {
          "e164": {
            "type": "string",
            "title": "E164"
          },
          "international": {
            "type": "string",
            "title": "International"
          },
          "national": {
            "type": "string",
            "title": "National"
          },
          "rfc3966": {
            "type": "string",
            "title": "Rfc3966"
          }
        },
        "type": "object",
        "required": [
          "e164",
          "international",
          "national",
          "rfc3966"
        ],
        "title": "FormatsInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationInfo": {
        "properties": {
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "country_name": {
            "type": "string",
            "title": "Country Name"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "timezones": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Timezones"
          }
        },
        "type": "object",
        "required": [
          "country_code",
          "country_name"
        ],
        "title": "LocationInfo"
      },
      "PhoneValidationResponse": {
        "properties": {
          "phone_input": {
            "type": "string",
            "title": "Phone Input"
          },
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid"
          },
          "is_possible": {
            "type": "boolean",
            "title": "Is Possible"
          },
          "formats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FormatsInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "carrier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CarrierInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "recommended_action": {
            "$ref": "#/components/schemas/RecommendedAction"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "phone_input",
          "is_valid",
          "is_possible",
          "risk_score",
          "risk_level",
          "recommended_action"
        ],
        "title": "PhoneValidationResponse"
      },
      "RecommendedAction": {
        "type": "string",
        "enum": [
          "ALLOW",
          "FLAG",
          "BLOCK"
        ],
        "title": "RecommendedAction"
      },
      "RiskLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "RiskLevel"
      },
      "TrustScoreRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Email address to evaluate (e.g. user@domain.com)"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "Phone number to evaluate (e.g. +14155552671)"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code",
            "description": "Default 2-letter ISO country code if phone number is local format",
            "default": "US"
          }
        },
        "type": "object",
        "title": "TrustScoreRequest"
      },
      "TrustScoreResponse": {
        "properties": {
          "trust_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Trust Score",
            "description": "Overall trust rating from 0 (fraudulent) to 100 (highly trusted)"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "description": "Composite risk rating from 0 (safe) to 100 (high risk)"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level",
            "description": "LOW, MEDIUM, or HIGH"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action",
            "description": "ALLOW, FLAG_FOR_REVIEW, or BLOCK"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signals",
            "description": "All contributing intelligence signals"
          },
          "email_intelligence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailVerificationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone_intelligence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhoneValidationResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "trust_score",
          "risk_score",
          "risk_level",
          "recommended_action"
        ],
        "title": "TrustScoreResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}